+++ requirements.txt
@@ -5,6 +5,7 @@ click==8.1.8
 colorama==0.4.6
 dnspython==2.7.0
 email_validator==2.2.0
+exceptiongroup==1.2.2
 fastapi==0.115.8
 fastapi-cli==0.0.7
 greenlet==3.1.1
