+++ setup.sh
@@ -1,0 +1,20 @@
+#!/bin/bash
+
+# Vai para o diretório do script
+cd "$(dirname "$0")"
+
+if [ -d "venv" ]; then
+    echo "Virtual environment already exists. Skipping creation."
+else
+    echo "Creating virtual environment..."
+    python3 -m venv venv
+fi
+
+echo "Activating virtual environment..."
+source venv/bin/activate
+
+echo "Installing dependencies..."
+pip install --upgrade pip
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+fi
